:label: update type annotation in sklean model
