+++ deepchem/models/models.py
@@ -2,27 +2,19 @@
 Contains an abstract base class that supports different ML models.
 """
 
-import sys
-import numpy as np
-import pandas as pd
-import joblib
 import os
 import shutil
 import tempfile
-import sklearn
 import logging
+from typing import Any, List, Optional, Sequence
+
+import numpy as np
 from sklearn.base import BaseEstimator
 
-import logging
-from deepchem.data import Dataset, pad_features
+from deepchem.data import Dataset
 from deepchem.metrics import Metric
 from deepchem.trans import Transformer, undo_transforms
-from deepchem.utils.save import load_from_disk
-from deepchem.utils.save import save_to_disk
 from deepchem.utils.evaluate import Evaluator
-
-from typing import Any, Dict, List, Optional, Sequence
-from deepchem.utils.typing import OneOrMany
 
 logger = logging.getLogger(__name__)
 
@@ -51,8 +43,8 @@ class Model(BaseEstimator):
     """
     if self.__class__.__name__ == "Model":
       raise ValueError(
-          "This constructor is for an abstract class and should never be called directly. Can only call from subclass constructors."
-      )
+          "This constructor is for an abstract class and should never be called directly."
+          "Can only call from subclass constructors.")
     self.model_dir_is_temp = False
     if model_dir is not None:
       if not os.path.exists(model_dir):
@@ -68,21 +60,17 @@ class Model(BaseEstimator):
     if 'model_dir_is_temp' in dir(self) and self.model_dir_is_temp:
       shutil.rmtree(self.model_dir)
 
-  def fit_on_batch(self, X: Sequence, y: Sequence, w: Sequence) -> float:
+  def fit_on_batch(self, X: Sequence, y: Sequence, w: Sequence):
     """Perform a single step of training.
 
     Parameters
     ----------
-    X: ndarray
+    X: np.ndarray
       the inputs for the batch
-    y: ndarray
+    y: np.ndarray
       the labels for the batch
-    w: ndarray
+    w: np.ndarray
       the weights for the batch
-
-    Returns
-    -------
-    the loss on the batch
     """
     raise NotImplementedError(
         "Each model is responsible for its own fit_on_batch method.")
@@ -140,7 +128,8 @@ class Model(BaseEstimator):
 
     Returns
     -------
-    The average loss over the most recent checkpoint interval. 
+    float
+      The average loss over the most recent checkpoint interval.
     """
     for epoch in range(nb_epoch):
       logger.info("Starting epoch %s" % str(epoch + 1))
@@ -152,28 +141,24 @@ class Model(BaseEstimator):
     return np.array(losses).mean()
 
   def predict(self, dataset: Dataset,
-              transformers: List[Transformer] = []) -> OneOrMany[np.ndarray]:
+              transformers: List[Transformer] = []) -> np.ndarray:
     """
     Uses self to make predictions on provided Dataset object.
 
-
     Parameters
     ----------
-    dataset: dc.data.Dataset
+    dataset: Dataset
       Dataset to make prediction on
-    transformers: list of dc.trans.Transformers
-      Transformers that the input data has been transformed by.  The output
+    transformers: List[Transformer]
+      Transformers that the input data has been transformed by. The output
       is passed through these transformers to undo the transformations.
 
     Returns
     -------
-    a NumPy array of the model produces a single output, or a list of arrays
-    if it produces multiple outputs
+    np.ndarray
+      A numpy array of predictions the model produces.
     """
     y_preds = []
-    n_tasks = self.get_num_tasks()
-    ind = 0
-
     for (X_batch, _, _, ids_batch) in dataset.iterbatches(deterministic=True):
       n_samples = len(X_batch)
       y_pred_batch = self.predict_on_batch(X_batch)
@@ -205,9 +190,9 @@ class Model(BaseEstimator):
 
     Parameters
     ----------
-    dataset: `dc.data.Dataset`
+    dataset: Dataset
       Dataset object.
-    metrics: dc.metrics.Metric/list[dc.metrics.Metric]/function
+    metrics: Metric / List[Metric] / function
       The set of metrics provided. This class attempts to do some
       intelligent handling of input. If a single `dc.metrics.Metric`
       object is provided or a list is provided, it will evaluate
@@ -218,11 +203,11 @@ class Model(BaseEstimator):
       `np.ndarray` objects and return a floating point score. The
       metric function may also accept a keyword argument
       `sample_weight` to account for per-sample weights.
-    transformers: list
+    transformers: List[Transformer]
       List of `dc.trans.Transformer` objects. These transformations
       must have been applied to `dataset` previously. The dataset will
       be untransformed for metric evaluation.
-    per_task_metrics: bool, optional
+    per_task_metrics: bool, optional (default False)
       If true, return computed metric for each task on multitask dataset.
     use_sample_weights: bool, optional (default False)
       If set, use per-sample weights `w`.
